chore: update v2 events references
