+++ src/events/birth/forms/correctionForm/requester.ts
@@ -18,8 +18,8 @@ import {
   not
 } from '@opencrvs/toolkit/events'
 import { InformantType, InformantTypeKey } from '../pages/informant'
-import { informantMessageDescriptors } from '@countryconfig/form/common/messages'
-import { IdType, idTypeOptions } from '@countryconfig/form/v2/person'
+import { informantMessageDescriptors } from '@countryconfig/events/common/messages'
+import { IdType, idTypeOptions } from '@countryconfig/events/person'
 import { nationalIdValidator } from '../../validators'
 
 const commonConfigs = {
