+++ src/events/birth/index.ts
@@ -32,8 +32,8 @@ import { BIRTH_CERTIFICATE_COLLECTOR_FORM } from './forms/printForm'
 import { PlaceOfBirth } from './forms/pages/child'
 import { CORRECTION_FORM } from './forms/correctionForm'
 import { dedupConfig } from './dedupConfig'
-import { BIRTH_LATE_REGISTRATION_TARGET_DAYS } from '@countryconfig/form/v2/constants'
-import { Event } from '@countryconfig/form/types'
+import { BIRTH_LATE_REGISTRATION_TARGET_DAYS } from '@countryconfig/events/constants'
+import { Event } from '@countryconfig/events/types'
 
 export const birthEvent = defineConfig({
   id: Event.Birth,
