+++ src/events/death/advancedSearch.ts
