+++ src/events/death/dedupConfig.ts
